# src/main/java/zone/gaygisiz/home/soft/banking/bank/entity/Bank.java
package zone.gaygisiz.home.soft.banking.bank.entity;

import lombok.Data;

import javax.persistence.*;

@Data
@Entity
public class Bank {

    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private Long id;

    private String name;

    @Column(unique = true)
    private String bic;

    @Column(unique = true)
    private String blz;
}